# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: B4P Youth API. Member routes require an active signed-in youth account; /admin routes require an active administrator. About and contact submission are open.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /me:
    get:
      operationId: getMyProfile
      tags: [profile]
      summary: Get the signed-in youth profile
      responses:
        "200":
          description: Current profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      operationId: updateMyProfile
      tags: [profile]
      summary: Update the signed-in youth profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
  /programs/{id}/applications:
    post:
      operationId: submitProgramApplication
      tags: [programs]
      summary: Apply to a published program
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgramApplicationInput"
      responses:
        "201":
          description: Submitted application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgramApplication"
        "409": { $ref: "#/components/responses/Conflict" }
...
  /events/{id}/registrations:
    post:
      operationId: registerForEvent
      tags: [events]
      summary: Register the signed-in youth for an event
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "201":
          description: Event registration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventRegistration"
        "409": { $ref: "#/components/responses/Conflict" }
...
  /me/notifications/{id}:
    patch:
      operationId: markMyNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationReadUpdate"
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema: { type: integer, minimum: 1 }
  responses:
...
    Conflict:
      description: Duplicate or conflicting request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    ApplicationStatus:
      type: string
      enum: [Submitted, Under Review, Accepted, Rejected]
...
    UserProfile:
      type: object
      required: [id, fullName, email, phone, gender, country, county, city, dateOfBirth, educationLevel, interests, imageUrl, active, createdAt]
      properties:
        id: { type: integer }
        fullName: { type: string }
        email: { type: string, format: email }
        phone: { type: string }
        gender: { type: string }
        country: { type: string }
        county: { type: string }
        city: { type: string }
        dateOfBirth: { type: string, format: date }
        educationLevel: { type: string }
        interests:
          type: array
          items: { type: string }
        imageUrl: { type: ["string", "null"], format: uri }
        active: { type: boolean }
        createdAt: { type: string, format: date-time }
    ProfileUpdate:
      type: object
      required: [fullName, phone, gender, country, county, city, dateOfBirth, educationLevel, interests]
      properties:
        fullName: { type: string, minLength: 1, maxLength: 160 }
        phone: { type: string, minLength: 1, maxLength: 40 }
        gender: { type: string, maxLength: 50 }
        country: { type: string, minLength: 1, maxLength: 100 }
        county: { type: string, maxLength: 100 }
        city: { type: string, maxLength: 100 }
        dateOfBirth: { type: string, format: date }
        educationLevel: { type: string, maxLength: 120 }
        interests:
          type: array
          items: { type: string, maxLength: 100 }
...
    ProgramApplicationInput:
      type: object
      required: [motivation]
      properties:
        motivation: { type: string, minLength: 20, maxLength: 5000 }
    ProgramApplication:
      type: object
      required: [id, programId, programName, status, motivation, adminNote, submittedAt]
      properties:
        id: { type: integer }
        programId: { type: integer }
        programName: { type: string }
        status: { $ref: "#/components/schemas/ApplicationStatus" }
        motivation: { type: string }
        adminNote: { type: ["string", "null"] }
        submittedAt: { type: string, format: date-time }
...
    EventRegistration:
      type: object
      required: [id, eventId, eventTitle, eventDate, registeredAt]
      properties:
        id: { type: integer }
        eventId: { type: integer }
        eventTitle: { type: string }
        eventDate: { type: string, format: date }
        registeredAt: { type: string, format: date-time }
...
    Notification:
      type: object
      required: [id, type, title, body, createdAt, readAt]
      properties:
        id: { type: integer }
        type: { type: string, enum: [opportunity, program, event, application, announcement] }
        title: { type: string }
        body: { type: string }
        createdAt: { type: string, format: date-time }
        readAt: { type: ["string", "null"], format: date-time }
    NotificationReadUpdate:
      type: object
      required: [read]
      properties:
        read: { type: boolean }
...